ui: components directories refactoring
