+++ core/ui/src/components/backup/RestoreSingleInstanceModal.vue
@@ -87,7 +87,7 @@
 <script>
 import { UtilService, TaskService, IconService } from "@nethserver/ns8-ui-lib";
 import to from "await-to-js";
-import NodeSelector from "@/components/NodeSelector";
+import NodeSelector from "@/components/misc/NodeSelector";
 import RestoreSingleInstanceSelector from "@/components/backup/RestoreSingleInstanceSelector";
 
 export default {
