+++ core/ui/src/components/domains/CreateDomainModal.vue
@@ -356,7 +356,7 @@ import {
   LottieService,
 } from "@nethserver/ns8-ui-lib";
 import to from "await-to-js";
-import NodeSelector from "@/components/NodeSelector";
+import NodeSelector from "@/components/misc/NodeSelector";
 
 export default {
   name: "CreateDomainModal",
