+++ core/ui/src/components/error/TaskHierarchy.vue
@@ -74,7 +74,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import "../styles/carbon-utils";
+@import "../../styles/carbon-utils";
 
 ul {
   padding: 0;
