+++ core/ui/src/components/misc/AreaChart.vue
