+++ core/ui/src/components/misc/NodeCard.vue
