+++ core/ui/src/components/misc/NodeSelector.vue
@@ -69,7 +69,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import "../styles/carbon-utils";
+@import "../../styles/carbon-utils";
 
 .node-id {
   margin-top: $spacing-05;
