+++ core/ui/src/components/misc/SkipRestoreAppsModal.vue
@@ -30,5 +30,5 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import "../styles/carbon-utils";
+@import "../../styles/carbon-utils";
 </style>
