+++ core/ui/src/components/shell/GlobalSearch.vue
@@ -330,7 +330,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import "../styles/carbon-utils";
+@import "../../styles/carbon-utils";
 
 .global-search {
   width: 50%;
@@ -440,7 +440,7 @@ export default {
 </style>
 
 <style lang="scss">
-@import "../styles/carbon-utils";
+@import "../../styles/carbon-utils";
 
 // global styles
