+++ core/ui/src/components/shell/SideMenuContent.vue
@@ -145,5 +145,5 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import "../styles/carbon-utils";
+@import "../../styles/carbon-utils";
 </style>
