+++ core/ui/src/components/software-center/ImageGallery.vue
@@ -81,7 +81,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import "../styles/carbon-utils";
+@import "../../styles/carbon-utils";
 
 .previews {
   display: flex;
