+++ core/ui/src/views/SoftwareCenterAppInstances.vue
@@ -283,7 +283,7 @@
 
 <script>
 import to from "await-to-js";
-import InstallAppModal from "../components/InstallAppModal";
+import InstallAppModal from "../components/software-center/InstallAppModal";
 import {
   QueryParamService,
   UtilService,
